:truck: Consolidate tests to mirror the package structure
